RDK-25250: Add --cert-status option to existing CURL commands RDKV.
Reason for change: add cert-status option to verify cert status.
Test Procedure: Test all the curl commands to establish secure TLS
connection.
Risks: Low.

(cherry picked from commit cb3da6e914c7070c5cc32edb1152d0bd7cd6b89f)
(cherry picked from commit d2ef2af939d399ce1f03b711b8cabb59b6fbe153)
(cherry picked from commit e63553b03dd057de96a8d9293e941ff1bd89421b)

+++ RFCbase.sh
@@ -746,6 +746,8 @@ sendHttpRequestToServer()
     FILENAME=$1
     URL=$2
     TryWithCodeBig=$3
+    EnableOCSPStapling="/tmp/.EnableOCSPStapling"
+    EnableOCSP="/tmp/.EnableOCSPCA"
 
     #Create json string
     if [ "$DEVICE_TYPE" = "broadband" ]; then 
@@ -790,11 +792,19 @@ sendHttpRequestToServer()
         eval $SIGN_CMD > /tmp/.signedRequest
         CB_SIGNED_REQUEST=`cat /tmp/.signedRequest`
         rm -f /tmp/.signedRequest
-        CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header"  "$IF_FLAG" --connect-timeout $timeout -m $timeout "$TLSFLAG" -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" \"$CB_SIGNED_REQUEST\""
+        if [ -f $EnableOCSPStapling ] || [ -f $EnableOCSP ]; then
+            CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header"  "$IF_FLAG" --cert-status --connect-timeout $timeout -m $timeout "$TLSFLAG" -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" \"$CB_SIGNED_REQUEST\""
+        else
+            CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header"  "$IF_FLAG" --connect-timeout $timeout -m $timeout "$TLSFLAG" -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" \"$CB_SIGNED_REQUEST\""
+        fi
         CURL_CMD_LOG=`echo "$CURL_CMD" | sed -ne 's#oauth_consumer_key=.*oauth_signature.*#-- <hidden> --#p'`
         rfcLogging "CURL_CMD: $CURL_CMD_LOG"
     else
-        CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header" "$IF_FLAG" --connect-timeout $timeout -m $timeout "$TLSFLAG"  -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" '$URL$JSONSTR'"
+        if [ -f $EnableOCSPStapling ] || [ -f $EnableOCSP ]; then
+            CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header" "$IF_FLAG" --cert-status --connect-timeout $timeout -m $timeout "$TLSFLAG"  -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" '$URL$JSONSTR'"
+        else
+            CURL_CMD="curl -w '%{http_code}\n'  -D "/tmp/curl_header" "$IF_FLAG" --connect-timeout $timeout -m $timeout "$TLSFLAG"  -H "configsethash:$valueHash" -H "configsettime:$valueTime" -o  \"$FILENAME\" '$URL$JSONSTR'"
+        fi
         rfcLogging "CURL_CMD: $CURL_CMD"
     fi
